# docs/api-contract.yaml
openapi: 3.0.3
info:
  title: EnjaRole API
  description: Character-based social network API
  version: 1.0.0

servers:
  - url: http://localhost:3000/api
    description: Local development

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    Character:
      type: object
      properties:
        id: { type: string, format: uuid }
        username: { type: string }
        displayName: { type: string }
        avatarUrl: { type: string, nullable: true }
        bio: { type: string, nullable: true }
        backstory: { type: string, nullable: true }
        personalityTraits: { type: array, items: { type: string } }
        createdAt: { type: string, format: date-time }

    AuthTokens:
      type: object
      properties:
        accessToken: { type: string }
        refreshToken: { type: string }
        character: { $ref: '#/components/schemas/Character' }

    Post:
      type: object
      properties:
        id: { type: string, format: uuid }
        content: { type: string }
        mediaUrls: { type: array, items: { type: string } }
        likeCount: { type: integer }
        commentCount: { type: integer }
        isLiked: { type: boolean }
        character: { $ref: '#/components/schemas/Character' }
        createdAt: { type: string, format: date-time }

    Comment:
      type: object
      properties:
        id: { type: string, format: uuid }
        content: { type: string }
        character: { $ref: '#/components/schemas/Character' }
        createdAt: { type: string, format: date-time }

    Story:
      type: object
      properties:
        id: { type: string, format: uuid }
        mediaUrl: { type: string }
        mediaType: { type: string, enum: [image] }
        character: { $ref: '#/components/schemas/Character' }
        expiresAt: { type: string, format: date-time }
        createdAt: { type: string, format: date-time }

    Message:
      type: object
      properties:
        id: { type: string, format: uuid }
        content: { type: string }
        sender: { $ref: '#/components/schemas/Character' }
        isRead: { type: boolean }
        createdAt: { type: string, format: date-time }

    Conversation:
      type: object
      properties:
        id: { type: string, format: uuid }
        participants: { type: array, items: { $ref: '#/components/schemas/Character' } }
        lastMessage: { $ref: '#/components/schemas/Message' }
        unreadCount: { type: integer }

    Notification:
      type: object
      properties:
        id: { type: string, format: uuid }
        type: { type: string, enum: [like, comment, follow, message] }
        payload: { type: object }
        isRead: { type: boolean }
        createdAt: { type: string, format: date-time }

paths:
  /auth/register:
    post:
      tags: [Auth]
      summary: Register new character account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, username, displayName]
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 8 }
                username: { type: string }
                displayName: { type: string }
                bio: { type: string }
                backstory: { type: string }
                personalityTraits: { type: array, items: { type: string } }
      responses:
        '201':
          description: Character created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthTokens' }

  /auth/login:
    post:
      tags: [Auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string }
                password: { type: string }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthTokens' }

  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken: { type: string }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthTokens' }

  /characters/me:
    get:
      tags: [Characters]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Character' }
    patch:
      tags: [Characters]
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                displayName: { type: string }
                bio: { type: string }
                backstory: { type: string }
                avatarUrl: { type: string }
                personalityTraits: { type: array, items: { type: string } }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Character' }

  /characters/{username}:
    get:
      tags: [Characters]
      parameters:
        - name: username
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: '#/components/schemas/Character' }
                  - type: object
                    properties:
                      postCount: { type: integer }
                      followerCount: { type: integer }
                      followingCount: { type: integer }
                      isFollowing: { type: boolean }

  /posts:
    post:
      tags: [Posts]
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                content: { type: string }
                mediaUrls: { type: array, items: { type: string } }
      responses:
        '201':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Post' }

  /feed:
    get:
      tags: [Social]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: mode
          in: query
          schema: { type: string, enum: [following, explore], default: following }
        - name: cursor
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/Post' } }
                  nextCursor: { type: string, nullable: true }

  /posts/{id}/like:
    post:
      tags: [Posts]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  liked: { type: boolean }
                  likeCount: { type: integer }

  /posts/{id}/comments:
    post:
      tags: [Posts]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                content: { type: string }
      responses:
        '201':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Comment' }

  /follow/{username}:
    post:
      tags: [Social]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: username
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  following: { type: boolean }

  /stories:
    post:
      tags: [Stories]
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                mediaUrl: { type: string }
                mediaType: { type: string, enum: [image] }
      responses:
        '201':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Story' }

  /stories/feed:
    get:
      tags: [Stories]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    character: { $ref: '#/components/schemas/Character' }
                    stories: { type: array, items: { $ref: '#/components/schemas/Story' } }

  /conversations:
    get:
      tags: [Messaging]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Conversation' }
    post:
      tags: [Messaging]
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username: { type: string }
      responses:
        '201':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Conversation' }

  /conversations/{id}/messages:
    get:
      tags: [Messaging]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
        - name: cursor
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/Message' } }
                  nextCursor: { type: string, nullable: true }
    post:
      tags: [Messaging]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                content: { type: string }
      responses:
        '201':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Message' }

  /notifications:
    get:
      tags: [Notifications]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: cursor
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/Notification' } }
                  nextCursor: { type: string, nullable: true }
                  unreadCount: { type: integer }

  /notifications/read-all:
    patch:
      tags: [Notifications]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }

  /media/upload:
    post:
      tags: [Media]
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                filename: { type: string }
                contentType: { type: string }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  uploadUrl: { type: string }
                  publicUrl: { type: string }